introduction de socket.io pour la communicatioon en temps reel avec le frontend

+++ models/resume.js
@@ -9,7 +9,7 @@ module.exports = (sequelize) => {
       allowNull: false,
     },
     userId: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(255),
       allowNull: false,
     },
     resume: {
